requirements.txt
rich>=13.0.0
prompt_toolkit>=3.0.0
pyyaml>=6.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-groq>=0.2.0
langchain-ollama>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
mcp>=1.0.0
